make vertex formats more dynamic, move away from floats, make texture loading way faster

+++ Rosella/src/main/kotlin/me/hydos/rosella/render/model/GuiRenderObject.kt
@@ -44,27 +44,28 @@ open class GuiRenderObject(
 
 		colour = Vector3f(0f, 0f, 0f)
 
+		// TODO: is this conversion doing what it should be? should convert int representing unsigned byte to signed byte through wrapping
 		renderInfo.consumer
 			.pos(-0.5f, -0.5f, 0f)
-			.color(colour.x().toInt(), colour.y().toInt(), colour.z().toInt())
+			.color(colour.x().toInt().toByte(), colour.y().toInt().toByte(), colour.z().toInt().toByte())
 			.uv(0f, 0f)
 			.nextVertex()
 
 		renderInfo.consumer
 			.pos(0.5f, -0.5f, 0f)
-			.color(colour.x().toInt(), colour.y().toInt(), colour.z().toInt())
+			.color(colour.x().toInt().toByte(), colour.y().toInt().toByte(), colour.z().toInt().toByte())
 			.uv(1f, 0f)
 			.nextVertex()
 
 		renderInfo.consumer
 			.pos(0.5f, 0.5f, 0f)
-			.color(colour.x().toInt(), colour.y().toInt(), colour.z().toInt())
+			.color(colour.x().toInt().toByte(), colour.y().toInt().toByte(), colour.z().toInt().toByte())
 			.uv(1f, 1f)
 			.nextVertex()
 
 		renderInfo.consumer
 			.pos(-0.5f, 0.5f, 0f)
-			.color(colour.x().toInt(), colour.y().toInt(), colour.z().toInt())
+			.color(colour.x().toInt().toByte(), colour.y().toInt().toByte(), colour.z().toInt().toByte())
 			.uv(0f, 1f)
 			.nextVertex()
